+++ src/iotvm-eventengine/src/main/java/org/softwareforce/iotvm/eventengine/cep/util/ValidationUtils.java
@@ -1,0 +1,16 @@
+package org.softwareforce.iotvm.eventengine.cep.util;
+
+/**
+ * TODO Docs.
+ *
+ * @author Dimitris Gkoulis
+ * @createdAt Wednesday 13 March 2024
+ */
+public final class ValidationUtils {
+
+  private ValidationUtils() {}
+
+  public static void assertIsTrue(final boolean bool, final String message) {
+    // TODO Implement.
+  }
+}